# docks/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service API",
    "description": "API для управления сервисами",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/appointments/{id}": {
      "put": {
        "tags": [
          "Appointments"
        ],
        "summary": "Обновить запись",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/appointments": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Создать новую запись",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/appointments/psychologist/{id}": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Получить записи по психологу",
        "operationId": "getByPsychologist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/client/{id}": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Получить записи по клиенту",
        "operationId": "getByClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppointmentUpdateRequest": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "example": "14:30:00"
          },
          "endTime": {
            "type": "string",
            "example": "14:30:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "BOOKED",
              "CONFIRMED",
              "CANCELLED"
            ]
          }
        }
      },
      "AppointmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "example": "14:30:00"
          },
          "endTime": {
            "type": "string",
            "example": "14:30:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "BOOKED",
              "CONFIRMED",
              "CANCELLED"
            ]
          },
          "psychologistId": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          }
        },
        "required": [
          "clientId",
          "date",
          "endTime",
          "id",
          "psychologistId",
          "startTime",
          "status"
        ]
      },
...
      "AppointmentRequest": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "example": "14:30:00"
          },
          "endTime": {
            "type": "string",
            "example": "14:30:00"
          },
          "psychologistId": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          }
        },
        "required": [
          "clientId",
          "date",
          "endTime",
          "psychologistId",
          "startTime"
        ]
      },
...
    }
  }
}